+++ docs/example/calendar/base.vue
@@ -1,0 +1,5 @@
+<template>
+
+  <ce-calendar></ce-calendar>
+
+</template>