+++ docs/src/router/zh-CN.ts
@@ -83,6 +83,11 @@ const zhCN: RouteRecordRaw[] = [
           path: "/zh-CN/components/dialog-selector",
           component: () => import("../../docs/zh-CN/components/dialog-selector.md"),
           meta: { title: "dialog-selector" }
+        },
+        {
+          path: "/zh-CN/components/calendar",
+          component: () => import("../../docs/zh-CN/components/calendar.md"),
+          meta: { title: "calendar" }
         }
     ]
   }