+++ packages/components/calendar/index.ts
@@ -1,0 +1,9 @@
+import { withInstall } from '@composite-ware/utils'
+
+export * from './src'
+
+import Calendar from './src/calendar.vue'
+
+export const CeCalendar = withInstall(Calendar)
+
+export default CeCalendar