+++ packages/components/calendar/src/index.ts
@@ -1,0 +1,8 @@
+import type { IDefineProps } from '../../types/index'
+import { withInstall } from '@composite-ware/utils'
+import Calendar from './calendar.vue'
+
+// import { } from './props'
+
+export const CeCalendar = withInstall(Calendar)
+// export type ITitleProps = IDefineProps<typeof TitleProps>