+++ CHANGELOG.md
@@ -1,1 +1,32 @@
-# Changelog
+---
+title: 更新日志
+meta:
+  - name: changeLog
+    content: composite-ware组件更新日志
+---
+## 0.0.7
+
+### 功能
+
+- 添加了日历组件 [canendar](http://10.28.89.11:9209/zh-CN/components/calendar)
+- 添加了 base-selector 的搜索条件请求接口的功能 | type=select，dialog-selector同
+- 添加了 base-selector 的搜索label 的 placeholder 选项 ，dialog-selector同[查看](http://10.28.89.11:9209/zh-CN/components/base-selector)
+- 补全了 base-selector 和 dialog-selector 的组件文档
+- 添加了文档项目的自动登入功能，便于查看需要请求的组件文档
+- 添加了组件项目的命令行功能 `pnpm make | npm run make` ，自动生成开发组件的所有文件并添加组件中，便于更快速的开发组件
+- 添加了重置文档项目的路由功能 `pnpm reset | npm run reset`，根据当前的组件，自动重置文档项目的路由
+
+### 修复
+
+- 修改了表格组件属性传输的错误
+
+## 0.0.6
+
+- 更新了以下组件和部分组件文档
+- [link](http://10.28.89.11:9209/zh-CN/components/link)
+- [card](http://10.28.89.11:9209/zh-CN/components/card)
+- [item](http://10.28.89.11:9209/zh-CN/components/item)
+- [title](http://10.28.89.11:9209/zh-CN/components/title)
+- [table](http://10.28.89.11:9209/zh-CN/components/table)
+- [base-selector](http://10.28.89.11:9209/zh-CN/components/base-selector)
+- [dialog-selector](http://10.28.89.11:9209/zh-CN/components/dialog-selector)